[DEV] Help Command Refactor (#39)

+++ src/commands/misc/help.ts
@@ -6,85 +6,99 @@ import {
   RuleBuilder,
 } from '@typeit/discord';
 import { environment } from '../../utils/environment';
+import { commandOverrides, commandTypes } from '../../data/help';
+import Utility from '../../utils/utility';
+import { Message, MessageEmbed } from 'discord.js';
+import { CommandType } from '../../types/help';
 
-const EXCLUDE_COMMANDS = ['help', 'purge', 'question'];
 export class Help {
   /**
-   * Init
+   * Create base message
+   * @param command
+   * @returns MessageEmbed
+   */
+  private createBaseMessage(
+    command: CommandMessage,
+    type?: CommandType
+  ): MessageEmbed {
+    const t = type ? ` ${Utility.captaliseFirstLetter(type)}` : '';
+    return new MessageEmbed()
+      .setColor(10181046)
+      .setAuthor(
+        `${command?.client?.user?.username}${t} Plugin Commands`,
+        command?.client?.user?.displayAvatarURL()
+      )
+      .setThumbnail('https://i.imgur.com/6CKmCtO.png');
+  }
+
+  /**
+   * Create help status message
    */
   private async createHelpStatus(
     command: CommandMessage,
-    allCommands: CommandInfos<any, RuleBuilder>[]
-  ): Promise<void> {
-    const fields = allCommands
-      .filter((c) => !EXCLUDE_COMMANDS.find((name) => name === c.commandName))
-      .map((c) => {
-        if (c.commandName === 'playerchoice') {
-          c.commandName =
-            'playerchoice <...@user(optional)>\nExclude users example: @user @user (add after command name)';
-        }
-
-        if (c.commandName === 'insult') {
-          c.commandName = 'insult <@user(optional)>';
-        }
-
-        if (c.commandName === 'compliment') {
-          c.commandName = 'compliment <@user(optional)>';
-        }
-
-        if (c.commandName === 'sayIt') {
-          c.commandName = 'sayIt <@user(optional)>';
-        }
-
-        if (c.commandName === 'urban') {
-          c.commandName = 'urban <word>';
-        }
-
-        if (c.commandName === 'poll') {
-          c.commandName = 'poll [question] [answer1] [answer2] ...';
-        }
-
-        if (c.commandName === 'weather') {
-          c.commandName = 'weather <area>';
-        }
-
-        if (c.commandName === 'twitch') {
-          c.commandName = 'twitch <username>';
-        }
-
-        if (c.commandName === 'steam') {
-          c.commandName = 'steam <vanity url name>';
-        }
-
-        if (c.commandName === 'instagram') {
-          c.commandName = 'instagram <username>';
-        }
-
-        if (c.commandName === 'minecraft') {
-          c.commandName = 'minecraft <ip>:<port(optional)>';
-        }
-
-        if (c.commandName === 'set minecraft') {
-          c.commandName = 'set minecraft <ip>:<port(optional)>';
-        }
-
+    allCommands: CommandInfos<any, RuleBuilder>[],
+    type?: CommandType
+  ): Promise<Message | void> {
+    if (type) {
+      const fields = allCommands.map((c) => {
+        const item = commandOverrides.find((cd) => cd.name === c.commandName);
+        c.commandName = item ? item.fullCommand : c.commandName;
         return {
           name: `**${c.description}**`,
           value: `\`@${environment.botName} ${c.commandName}\``,
         };
       });
 
-    command.channel.send({
-      embed: {
-        color: 10181046,
-        author: {
-          name: `${command?.client?.user?.username} Plugin Commands`,
-          icon_url: command?.client?.user?.displayAvatarURL(),
+      const message = this.createBaseMessage(command, type);
+      message.addFields([...fields]);
+
+      await command.delete();
+      return command.channel.send(message);
+    } else {
+      const message = this.createBaseMessage(command);
+      message.addFields([
+        {
+          name: `Fun`,
+          value: `\`@${command?.client?.user?.username} help fun\``,
         },
-        fields,
-      },
-    });
-    return Promise.resolve();
+        {
+          name: `Games`,
+          value: `\`@${command?.client?.user?.username} help games\``,
+        },
+        {
+          name: `Searchers`,
+          value: `\`@${command?.client?.user?.username} help searchers\``,
+        },
+      ]);
+
+      if (Utility.isAdmin(command)) {
+        message.addField(
+          'Admin',
+          `\`@${command?.client?.user?.username} help admin\``
+        );
+      }
+      await command.delete();
+      return command.channel.send(message);
+    }
+  }
+
+  /**
+   * Fetch Commands
+   * @param type
+   */
+  fetchCommands(type?: CommandType): CommandInfos<any, RuleBuilder>[] {
+    const commands = Client.getCommands();
+
+    if (type) {
+      const commandNames = commandOverrides
+        .filter((c) => c.type === type)
+        .map((c) => c.name);
+      return commands.filter(
+        (c) => commandNames.indexOf(c.commandName as string) > -1
+      );
+    }
+
+    return commands.filter((c) => c.commandName !== 'help');
   }
 
   /**
@@ -94,10 +108,31 @@ export class Help {
    * @returns
    */
   @Command('help')
-  helpInit(command: CommandMessage): Promise<void> {
-    const allCommands = Client.getCommands();
-    command.delete();
-    return this.createHelpStatus(command, allCommands).catch(() => {
+  async helpInit(command: CommandMessage): Promise<Message | void> {
+    const type = Utility.getOptionFromCommand(
+      command.content,
+      2,
+      ' '
+    ) as CommandType;
+
+    if (type && !commandTypes.find((c) => c === type)) {
+      await command.delete();
+      return command.channel
+        .send(
+          `**This command grouping does not exist! Please use just help to see valid groupings.**`
+        )
+        .then((m) => m.delete({ timeout: 5000 }));
+    }
+
+    if (!Utility.isAdmin(command)) {
+      await command.delete();
+      return command.channel
+        .send(`**You don't have the permissions for this**`)
+        .then((m) => m.delete({ timeout: 5000 }));
+    }
+
+    const allCommands = this.fetchCommands(type);
+    return this.createHelpStatus(command, allCommands, type).catch(() => {
       command.reply(environment.error);
     });
   }
